Add readonly and move app to a folder (#17)
* add readonly and move app to a folder

* add utilities files in store

* add app folder under src
