+++ src/app/redux/constants.ts
@@ -1,0 +1,1 @@
+export const UPDATE_SELECTED_STOCK = "UPDATE_SELECTED_STOCK";
