+++ src/store/actions.ts
@@ -1,11 +1,7 @@
 import { BOOTSTRAP, RESET } from "./constants";
+import { Action } from "./utilities";
 
 export const bootstrap = (): BootstrapAction => ({ type: BOOTSTRAP });
 export type BootstrapAction = Action<typeof BOOTSTRAP>;
 export const resetAction = () => ({ type: RESET });
 export type ResetAction = Action<typeof RESET>;
-
-export type Action<T extends string> = { type: T };
-export type ActionWithPayload<T extends string, P> = Action<T> & {
-  payload: P;
-};
