+++ src/store/index.ts
@@ -1,2 +1,3 @@
 export * from "./constants";
 export * from "./reducer";
+export * from "./utilities";
