+++ src/store/utilities.ts
@@ -1,0 +1,4 @@
+export type Action<T extends string> = { type: T };
+export type ActionWithPayload<T extends string, P> = Action<T> & {
+  payload: P;
+};
